Create an Application

This function creates an Application

createApplication(
  SchoolYearID = NULL,
  DistrictID = NULL,
  SignatureDate = NULL,
  DistrictReceivedDate = NULL,
  EffectiveDate = NULL,
  HouseholdSize = NULL,
  Status = NULL,
  ApplicationNumber = NULL,
  EntryMethod = NULL,
  AssistanceSourceID = NULL,
  CaseNumber = NULL,
  NoSSN = NULL,
  LastFourSSN = NULL,
  EligibilityCategoryID = NULL,
  IsDenied = NULL,
  Comment = NULL,
  ExcludeRandomSelectionReasonID = NULL,
  SelectedForVerification = NULL,
  DirectVerificationComplete = NULL,
  PreviouslySelectedForVerification = NULL,
  FirstRequestSent = NULL,
  SecondRequestSent = NULL,
  NoResponse = NULL,
  DirectVerificationCompletionDate = NULL,
  FirstRequestNoticeDate = NULL,
  FirstRequestExpectedDate = NULL,
  SecondRequestNoticeDate = NULL,
  SecondRequestExpectedDate = NULL,
  VerificationResult = NULL,
  NoticeType = NULL,
  StatusChangeEffectiveDate = NULL,
  UserIDVerifiedBy = NULL,
  VerificationDate = NULL,
  ReasonForChangeIncome = NULL,
  ReasonForChangeHouseholdSize = NULL,
  ReasonForChangeOther = NULL,
  NameIDDeterminedBy = NULL,
  DeterminedDate = NULL,
  SelectedChangeReasonIncome = NULL,
  SelectedChangeReasonHouseHoldSize = NULL,
  SelectedChangeReasonDirectlyVerified = NULL,
  SelectedChangeReasonNoResponse = NULL,
  SelectedChangeReasonOther = NULL,
  IsSelectedForVerificationAsCaseNumber = NULL,
  VerifiedForCauseReason = NULL,
  VerifiedForCause = NULL,
  TotalChildIncomeAmount = NULL,
  TotalChildFrequency = NULL,
  NoChildIncome = NULL,
  ConfirmationReviewReplaceApplication = NULL,
  BenefitLevelChangeDate = NULL,
  VerificationConfirmationDate = NULL,
  UserIDConfirmedBy = NULL,
  ConfirmationStatus = NULL,
  ProgramOfEligibility = NULL,
  MealCode = NULL,
  AcceptFeeManagementOptionalBenefit = NULL,
  OnlineApplicationID = NULL,
  ApplicationFormatID = NULL,
  entityId = 1,
  schoolYearId = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the created Application. Each defaults to NULL.

Value

A newly created Application Skyward's Knowledge Hub