Create a BankAccount

This function creates a BankAccount

createBankAccount(
  DistrictID = NULL,
  Code = NULL,
  Description = NULL,
  Type = NULL,
  MaskIDCash = NULL,
  MaskIDDueTo = NULL,
  MaskIDDueFrom = NULL,
  NextCheckNumber = NULL,
  NextWireTransferNumber = NULL,
  NextACHNumber = NULL,
  AccountType = NULL,
  RoutingNumber = NULL,
  AccountNumber = NULL,
  ACHCompanyName = NULL,
  ACHCompanyIdentification = NULL,
  ACHImmediateOriginName = NULL,
  ACHImmediateOrigin = NULL,
  ACHImmediateDestinationName = NULL,
  ACHImmediateDestination = NULL,
  ACHOriginatingDFIIdentification = NULL,
  ACHFileIDModifier = NULL,
  ACHUserDefinedHeader = NULL,
  ACHSettlementDate = NULL,
  ACHServiceClass = NULL,
  ACHHasCarriageReturnLineFeed = NULL,
  ACHHasOffsetDebit = NULL,
  PrintedEmployerInfo1 = NULL,
  PrintedEmployerInfo2 = NULL,
  PrintedEmployerInfo3 = NULL,
  PrintedEmployerInfo4 = NULL,
  PrintedEmployerInfo5 = NULL,
  PrintedBankInfo1 = NULL,
  PrintedBankInfo2 = NULL,
  PrintedBankInfo3 = NULL,
  PrintedBankInfo4 = NULL,
  PrintedBankInfo5 = NULL,
  PrintedCheckNumberLabel = NULL,
  PrintedDepositNumberLabel = NULL,
  PrintedMessageAboveSignature = NULL,
  PrintSocialSecurityNumber = NULL,
  PrintContractInformation = NULL,
  PrintCheckLocationOnCheck = NULL,
  PrintCheckLocationOnDeposit = NULL,
  PrintMICRLine = NULL,
  PrintBarCodeOnCheck = NULL,
  PrintBarCodeOnDeposit = NULL,
  PrintCheckNumberOnCheck = NULL,
  PrintCheckNumberOnDeposit = NULL,
  PrintTaxDeductions = NULL,
  ElectronicSignatureIDPayroll = NULL,
  ElectronicSignatureIDAccountsPayable = NULL,
  MediaIDEmployerInfoImage = NULL,
  MediaIDBankInfoImage = NULL,
  PrintedNotes = NULL,
  PrintedHorizontalAdjustment = NULL,
  PrintedVerticalAdjustment = NULL,
  EnvelopeHorizontalAdjustmentPayrollCheck = NULL,
  EnvelopeVerticalAdjustmentPayrollCheck = NULL,
  EnvelopeHorizontalAdjustmentPayrollACH = NULL,
  EnvelopeVerticalAdjustmentPayrollACH = NULL,
  EnvelopeHorizontalAdjustmentAccountsPayable = NULL,
  EnvelopeVerticalAdjustmentAccountsPayable = NULL,
  WrapInvoiceNumbers = NULL,
  WrapInvoiceDescriptions = NULL,
  PressureSealReturnAddressInfo1 = NULL,
  PressureSealReturnAddressInfo2 = NULL,
  PressureSealReturnAddressInfo3 = NULL,
  PressureSealReturnAddressInfo4 = NULL,
  PressureSealReturnAddressInfo5 = NULL,
  PressureSealReturnAddressInfo6 = NULL,
  PressureSealPostageInfo1 = NULL,
  PressureSealPostageInfo2 = NULL,
  PressureSealPostageInfo3 = NULL,
  PressureSealPostageInfo4 = NULL,
  PressureSealPostageInfo5 = NULL,
  AllowPayroll = NULL,
  AllowAccountsPayable = NULL,
  AllowCashReceipt = NULL,
  IsActive = NULL,
  AccountsPayableCheckStock = NULL,
  PayrollCheckStock = NULL,
  PayrollDepositStock = NULL,
  AddendumInvoiceDisplayOption = NULL,
  NextManualCheckNumber = NULL,
  PrintFundOnCheck = NULL,
  PrintFundOnDeposit = NULL,
  NextEpayableCheckNumber = NULL,
  ReportIDEpayableFile = NULL,
  AllowAccountsReceivable = NULL,
  RoutingNumberOverrideACH = NULL,
  AccountNumberOverrideACH = NULL,
  ActivityAccountingGroupID = NULL,
  PrintZeroValuePayrollChecks = NULL,
  PrintZeroValueAccountsPayableChecks = NULL,
  ACHIndividualIdentification = NULL,
  entityId = 1,
  schoolYearId = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the created BankAccount. Each defaults to NULL.

Value

A newly created BankAccount Skyward's Knowledge Hub