Get a CRDCSchoolDisciplineWithDisabilityV1

This function returns a dataframe or json object of a CRDCSchoolDisciplineWithDisabilityV1

getCRDCSchoolDisciplineWithDisabilityV1(
  CRDCSchoolDisciplineWithDisabilityV1ID,
  CorporalMaleHispanicCount = F,
  CorporalMaleAlaskanCount = F,
  CorporalMaleAsianCount = F,
  CorporalMaleHawaiianCount = F,
  CorporalMaleBlackCount = F,
  CorporalMaleWhiteCount = F,
  CorporalMaleMultiRaceCount = F,
  CorporalMaleLEPCount = F,
  CorporalMaleSection504Count = F,
  CorporalFemaleHispanicCount = F,
  CorporalFemaleAlaskanCount = F,
  CorporalFemaleAsianCount = F,
  CorporalFemaleHawaiianCount = F,
  CorporalFemaleBlackCount = F,
  CorporalFemaleWhiteCount = F,
  CorporalFemaleMultiRaceCount = F,
  CorporalFemaleLEPCount = F,
  CorporalFemaleSection504Count = F,
  InSchoolMaleHispanicCount = F,
  InSchoolMaleAlaskanCount = F,
  InSchoolMaleAsianCount = F,
  InSchoolMaleHawaiianCount = F,
  InSchoolMaleBlackCount = F,
  InSchoolMaleWhiteCount = F,
  InSchoolMaleMultiRaceCount = F,
  InSchoolMaleLEPCount = F,
  InSchoolMaleSection504Count = F,
  InSchoolFemaleHispanicCount = F,
  InSchoolFemaleAlaskanCount = F,
  InSchoolFemaleAsianCount = F,
  InSchoolFemaleHawaiianCount = F,
  InSchoolFemaleBlackCount = F,
  InSchoolFemaleWhiteCount = F,
  InSchoolFemaleMultiRaceCount = F,
  InSchoolFemaleLEPCount = F,
  InSchoolFemaleSection504Count = F,
  OneOutOfSchoolMaleHispanicCount = F,
  OneOutOfSchoolMaleAlaskanCount = F,
  OneOutOfSchoolMaleAsianCount = F,
  OneOutOfSchoolMaleHawaiianCount = F,
  OneOutOfSchoolMaleBlackCount = F,
  OneOutOfSchoolMaleWhiteCount = F,
  OneOutOfSchoolMaleMultiRaceCount = F,
  OneOutOfSchoolMaleLEPCount = F,
  OneOutOfSchoolMaleSection504Count = F,
  OneOutOfSchoolFemaleHispanicCount = F,
  OneOutOfSchoolFemaleAlaskanCount = F,
  OneOutOfSchoolFemaleAsianCount = F,
  OneOutOfSchoolFemaleHawaiianCount = F,
  OneOutOfSchoolFemaleBlackCount = F,
  OneOutOfSchoolFemaleWhiteCount = F,
  OneOutOfSchoolFemaleMultiRaceCount = F,
  OneOutOfSchoolFemaleLEPCount = F,
  OneOutOfSchoolFemaleSection504Count = F,
  MultipleOutOfSchoolMaleHispanicCount = F,
  MultipleOutOfSchoolMaleAlaskanCount = F,
  MultipleOutOfSchoolMaleAsianCount = F,
  MultipleOutOfSchoolMaleHawaiianCount = F,
  MultipleOutOfSchoolMaleBlackCount = F,
  MultipleOutOfSchoolMaleWhiteCount = F,
  MultipleOutOfSchoolMaleMultiRaceCount = F,
  MultipleOutOfSchoolMaleLEPCount = F,
  MultipleOutOfSchoolMaleSection504Count = F,
  MultipleOutOfSchoolFemaleHispanicCount = F,
  MultipleOutOfSchoolFemaleAlaskanCount = F,
  MultipleOutOfSchoolFemaleAsianCount = F,
  MultipleOutOfSchoolFemaleHawaiianCount = F,
  MultipleOutOfSchoolFemaleBlackCount = F,
  MultipleOutOfSchoolFemaleWhiteCount = F,
  MultipleOutOfSchoolFemaleMultiRaceCount = F,
  MultipleOutOfSchoolFemaleLEPCount = F,
  MultipleOutOfSchoolFemaleSection504Count = F,
  ExpulsionWithServicesMaleHispanicCount = F,
  ExpulsionWithServicesMaleAlaskanCount = F,
  ExpulsionWithServicesMaleAsianCount = F,
  ExpulsionWithServicesMaleHawaiianCount = F,
  ExpulsionWithServicesMaleBlackCount = F,
  ExpulsionWithServicesMaleWhiteCount = F,
  ExpulsionWithServicesMaleMultiRaceCount = F,
  ExpulsionWithServicesMaleLEPCount = F,
  ExpulsionWithServicesMaleSection504Count = F,
  ExpulsionWithServicesFemaleHispanicCount = F,
  ExpulsionWithServicesFemaleAlaskanCount = F,
  ExpulsionWithServicesFemaleAsianCount = F,
  ExpulsionWithServicesFemaleHawaiianCount = F,
  ExpulsionWithServicesFemaleBlackCount = F,
  ExpulsionWithServicesFemaleWhiteCount = F,
  ExpulsionWithServicesFemaleMultiRaceCount = F,
  ExpulsionWithServicesFemaleLEPCount = F,
  ExpulsionWithServicesFemaleSection504Count = F,
  ExpulsionNoServicesMaleHispanicCount = F,
  ExpulsionNoServicesMaleAlaskanCount = F,
  ExpulsionNoServicesMaleAsianCount = F,
  ExpulsionNoServicesMaleHawaiianCount = F,
  ExpulsionNoServicesMaleBlackCount = F,
  ExpulsionNoServicesMaleWhiteCount = F,
  ExpulsionNoServicesMaleMultiRaceCount = F,
  ExpulsionNoServicesMaleLEPCount = F,
  ExpulsionNoServicesMaleSection504Count = F,
  ExpulsionNoServicesFemaleHispanicCount = F,
  ExpulsionNoServicesFemaleAlaskanCount = F,
  ExpulsionNoServicesFemaleAsianCount = F,
  ExpulsionNoServicesFemaleHawaiianCount = F,
  ExpulsionNoServicesFemaleBlackCount = F,
  ExpulsionNoServicesFemaleWhiteCount = F,
  ExpulsionNoServicesFemaleMultiRaceCount = F,
  ExpulsionNoServicesFemaleLEPCount = F,
  ExpulsionNoServicesFemaleSection504Count = F,
  ZeroToleranceMaleHispanicCount = F,
  ZeroToleranceMaleAlaskanCount = F,
  ZeroToleranceMaleAsianCount = F,
  ZeroToleranceMaleHawaiianCount = F,
  ZeroToleranceMaleBlackCount = F,
  ZeroToleranceMaleWhiteCount = F,
  ZeroToleranceMaleMultiRaceCount = F,
  ZeroToleranceMaleLEPCount = F,
  ZeroToleranceMaleSection504Count = F,
  ZeroToleranceFemaleHispanicCount = F,
  ZeroToleranceFemaleAlaskanCount = F,
  ZeroToleranceFemaleAsianCount = F,
  ZeroToleranceFemaleHawaiianCount = F,
  ZeroToleranceFemaleBlackCount = F,
  ZeroToleranceFemaleWhiteCount = F,
  ZeroToleranceFemaleMultiRaceCount = F,
  ZeroToleranceFemaleLEPCount = F,
  ZeroToleranceFemaleSection504Count = F,
  LawEnforcementMaleHispanicCount = F,
  LawEnforcementMaleAlaskanCount = F,
  LawEnforcementMaleAsianCount = F,
  LawEnforcementMaleHawaiianCount = F,
  LawEnforcementMaleBlackCount = F,
  LawEnforcementMaleWhiteCount = F,
  LawEnforcementMaleMultiRaceCount = F,
  LawEnforcementMaleLEPCount = F,
  LawEnforcementMaleSection504Count = F,
  LawEnforcementFemaleHispanicCount = F,
  LawEnforcementFemaleAlaskanCount = F,
  LawEnforcementFemaleAsianCount = F,
  LawEnforcementFemaleHawaiianCount = F,
  LawEnforcementFemaleBlackCount = F,
  LawEnforcementFemaleWhiteCount = F,
  LawEnforcementFemaleMultiRaceCount = F,
  LawEnforcementFemaleLEPCount = F,
  LawEnforcementFemaleSection504Count = F,
  ArrestMaleHispanicCount = F,
  ArrestMaleAlaskanCount = F,
  ArrestMaleAsianCount = F,
  ArrestMaleHawaiianCount = F,
  ArrestMaleBlackCount = F,
  ArrestMaleWhiteCount = F,
  ArrestMaleMultiRaceCount = F,
  ArrestMaleLEPCount = F,
  ArrestMaleSection504Count = F,
  ArrestFemaleHispanicCount = F,
  ArrestFemaleAlaskanCount = F,
  ArrestFemaleAsianCount = F,
  ArrestFemaleHawaiianCount = F,
  ArrestFemaleBlackCount = F,
  ArrestFemaleWhiteCount = F,
  ArrestFemaleMultiRaceCount = F,
  ArrestFemaleLEPCount = F,
  ArrestFemaleSection504Count = F,
  CivilRightsDataCollectionID = F,
  CivilRightsDataCollectionRunHistoryID = F,
  UserIDCreatedBy = F,
  CreatedTime = F,
  UserIDModifiedBy = F,
  ModifiedTime = F,
  FederalNCESSchoolID = F,
  fieldPaths = NULL,
  entityId = 1,
  schoolYearId = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

CRDCSchoolDisciplineWithDisabilityV1ID
The ID of the CRDCSchoolDisciplineWithDisabilityV1 to return.

fieldPaths
Fields from other objects with 'Many to One' or 'One to One' relationships to the given object listed as text. Run getSchemaForObjects('CRDCSchoolDisciplineWithDisabilityV1') to get more field paths.

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
A TRUE or FALSE value determining whether or not to return the field for the given CRDCSchoolDisciplineWithDisabilityV1. Defaults to FALSE for all return fields which, for convenience, returns all fields for the CRDCSchoolDisciplineWithDisabilityV1.

Value

A dataframe or of CRDCSchoolDisciplineWithDisabilityV1 Skyward's Knowledge Hub