Get a CRDCSchoolRetentionCountV1

This function returns a dataframe or json object of a CRDCSchoolRetentionCountV1

getCRDCSchoolRetentionCountV1(
  CRDCSchoolRetentionCountV1ID,
  CivilRightsDataCollectionID = F,
  CivilRightsDataCollectionRunHistoryID = F,
  CreatedTime = F,
  FederalNCESSchoolID = F,
  Grade01FemaleAlaskanCount = F,
  Grade01FemaleAsianCount = F,
  Grade01FemaleBlackCount = F,
  Grade01FemaleHawaiianCount = F,
  Grade01FemaleHispanicCount = F,
  Grade01FemaleIDEACount = F,
  Grade01FemaleLEPCount = F,
  Grade01FemaleMultiRaceCount = F,
  Grade01FemaleSection504Count = F,
  Grade01FemaleWhiteCount = F,
  Grade01MaleAlaskanCount = F,
  Grade01MaleAsianCount = F,
  Grade01MaleBlackCount = F,
  Grade01MaleHawaiianCount = F,
  Grade01MaleHispanicCount = F,
  Grade01MaleIDEACount = F,
  Grade01MaleLEPCount = F,
  Grade01MaleMultiRaceCount = F,
  Grade01MaleSection504Count = F,
  Grade01MaleWhiteCount = F,
  Grade02FemaleAlaskanCount = F,
  Grade02FemaleAsianCount = F,
  Grade02FemaleBlackCount = F,
  Grade02FemaleHawaiianCount = F,
  Grade02FemaleHispanicCount = F,
  Grade02FemaleIDEACount = F,
  Grade02FemaleLEPCount = F,
  Grade02FemaleMultiRaceCount = F,
  Grade02FemaleSection504Count = F,
  Grade02FemaleWhiteCount = F,
  Grade02MaleAlaskanCount = F,
  Grade02MaleAsianCount = F,
  Grade02MaleBlackCount = F,
  Grade02MaleHawaiianCount = F,
  Grade02MaleHispanicCount = F,
  Grade02MaleIDEACount = F,
  Grade02MaleLEPCount = F,
  Grade02MaleMultiRaceCount = F,
  Grade02MaleSection504Count = F,
  Grade02MaleWhiteCount = F,
  Grade03FemaleAlaskanCount = F,
  Grade03FemaleAsianCount = F,
  Grade03FemaleBlackCount = F,
  Grade03FemaleHawaiianCount = F,
  Grade03FemaleHispanicCount = F,
  Grade03FemaleIDEACount = F,
  Grade03FemaleLEPCount = F,
  Grade03FemaleMultiRaceCount = F,
  Grade03FemaleSection504Count = F,
  Grade03FemaleWhiteCount = F,
  Grade03MaleAlaskanCount = F,
  Grade03MaleAsianCount = F,
  Grade03MaleBlackCount = F,
  Grade03MaleHawaiianCount = F,
  Grade03MaleHispanicCount = F,
  Grade03MaleIDEACount = F,
  Grade03MaleLEPCount = F,
  Grade03MaleMultiRaceCount = F,
  Grade03MaleSection504Count = F,
  Grade03MaleWhiteCount = F,
  Grade04FemaleAlaskanCount = F,
  Grade04FemaleAsianCount = F,
  Grade04FemaleBlackCount = F,
  Grade04FemaleHawaiianCount = F,
  Grade04FemaleHispanicCount = F,
  Grade04FemaleIDEACount = F,
  Grade04FemaleLEPCount = F,
  Grade04FemaleMultiRaceCount = F,
  Grade04FemaleSection504Count = F,
  Grade04FemaleWhiteCount = F,
  Grade04MaleAlaskanCount = F,
  Grade04MaleAsianCount = F,
  Grade04MaleBlackCount = F,
  Grade04MaleHawaiianCount = F,
  Grade04MaleHispanicCount = F,
  Grade04MaleIDEACount = F,
  Grade04MaleLEPCount = F,
  Grade04MaleMultiRaceCount = F,
  Grade04MaleSection504Count = F,
  Grade04MaleWhiteCount = F,
  Grade05FemaleAlaskanCount = F,
  Grade05FemaleAsianCount = F,
  Grade05FemaleBlackCount = F,
  Grade05FemaleHawaiianCount = F,
  Grade05FemaleHispanicCount = F,
  Grade05FemaleIDEACount = F,
  Grade05FemaleLEPCount = F,
  Grade05FemaleMultiRaceCount = F,
  Grade05FemaleSection504Count = F,
  Grade05FemaleWhiteCount = F,
  Grade05MaleAlaskanCount = F,
  Grade05MaleAsianCount = F,
  Grade05MaleBlackCount = F,
  Grade05MaleHawaiianCount = F,
  Grade05MaleHispanicCount = F,
  Grade05MaleIDEACount = F,
  Grade05MaleLEPCount = F,
  Grade05MaleMultiRaceCount = F,
  Grade05MaleSection504Count = F,
  Grade05MaleWhiteCount = F,
  Grade06FemaleAlaskanCount = F,
  Grade06FemaleAsianCount = F,
  Grade06FemaleBlackCount = F,
  Grade06FemaleHawaiianCount = F,
  Grade06FemaleHispanicCount = F,
  Grade06FemaleIDEACount = F,
  Grade06FemaleLEPCount = F,
  Grade06FemaleMultiRaceCount = F,
  Grade06FemaleSection504Count = F,
  Grade06FemaleWhiteCount = F,
  Grade06MaleAlaskanCount = F,
  Grade06MaleAsianCount = F,
  Grade06MaleBlackCount = F,
  Grade06MaleHawaiianCount = F,
  Grade06MaleHispanicCount = F,
  Grade06MaleIDEACount = F,
  Grade06MaleLEPCount = F,
  Grade06MaleMultiRaceCount = F,
  Grade06MaleSection504Count = F,
  Grade06MaleWhiteCount = F,
  Grade07FemaleAlaskanCount = F,
  Grade07FemaleAsianCount = F,
  Grade07FemaleBlackCount = F,
  Grade07FemaleHawaiianCount = F,
  Grade07FemaleHispanicCount = F,
  Grade07FemaleIDEACount = F,
  Grade07FemaleLEPCount = F,
  Grade07FemaleMultiRaceCount = F,
  Grade07FemaleSection504Count = F,
  Grade07FemaleWhiteCount = F,
  Grade07MaleAlaskanCount = F,
  Grade07MaleAsianCount = F,
  Grade07MaleBlackCount = F,
  Grade07MaleHawaiianCount = F,
  Grade07MaleHispanicCount = F,
  Grade07MaleIDEACount = F,
  Grade07MaleLEPCount = F,
  Grade07MaleMultiRaceCount = F,
  Grade07MaleSection504Count = F,
  Grade07MaleWhiteCount = F,
  Grade08FemaleAlaskanCount = F,
  Grade08FemaleAsianCount = F,
  Grade08FemaleBlackCount = F,
  Grade08FemaleHawaiianCount = F,
  Grade08FemaleHispanicCount = F,
  Grade08FemaleIDEACount = F,
  Grade08FemaleLEPCount = F,
  Grade08FemaleMultiRaceCount = F,
  Grade08FemaleSection504Count = F,
  Grade08FemaleWhiteCount = F,
  Grade08MaleAlaskanCount = F,
  Grade08MaleAsianCount = F,
  Grade08MaleBlackCount = F,
  Grade08MaleHawaiianCount = F,
  Grade08MaleHispanicCount = F,
  Grade08MaleIDEACount = F,
  Grade08MaleLEPCount = F,
  Grade08MaleMultiRaceCount = F,
  Grade08MaleSection504Count = F,
  Grade08MaleWhiteCount = F,
  Grade09FemaleAlaskanCount = F,
  Grade09FemaleAsianCount = F,
  Grade09FemaleBlackCount = F,
  Grade09FemaleHawaiianCount = F,
  Grade09FemaleHispanicCount = F,
  Grade09FemaleIDEACount = F,
  Grade09FemaleLEPCount = F,
  Grade09FemaleMultiRaceCount = F,
  Grade09FemaleSection504Count = F,
  Grade09FemaleWhiteCount = F,
  Grade09MaleAlaskanCount = F,
  Grade09MaleAsianCount = F,
  Grade09MaleBlackCount = F,
  Grade09MaleHawaiianCount = F,
  Grade09MaleHispanicCount = F,
  Grade09MaleIDEACount = F,
  Grade09MaleLEPCount = F,
  Grade09MaleMultiRaceCount = F,
  Grade09MaleSection504Count = F,
  Grade09MaleWhiteCount = F,
  Grade10FemaleAlaskanCount = F,
  Grade10FemaleAsianCount = F,
  Grade10FemaleBlackCount = F,
  Grade10FemaleHawaiianCount = F,
  Grade10FemaleHispanicCount = F,
  Grade10FemaleIDEACount = F,
  Grade10FemaleLEPCount = F,
  Grade10FemaleMultiRaceCount = F,
  Grade10FemaleSection504Count = F,
  Grade10FemaleWhiteCount = F,
  Grade10MaleAlaskanCount = F,
  Grade10MaleAsianCount = F,
  Grade10MaleBlackCount = F,
  Grade10MaleHawaiianCount = F,
  Grade10MaleHispanicCount = F,
  Grade10MaleIDEACount = F,
  Grade10MaleLEPCount = F,
  Grade10MaleMultiRaceCount = F,
  Grade10MaleSection504Count = F,
  Grade10MaleWhiteCount = F,
  Grade11FemaleAlaskanCount = F,
  Grade11FemaleAsianCount = F,
  Grade11FemaleBlackCount = F,
  Grade11FemaleHawaiianCount = F,
  Grade11FemaleHispanicCount = F,
  Grade11FemaleIDEACount = F,
  Grade11FemaleLEPCount = F,
  Grade11FemaleMultiRaceCount = F,
  Grade11FemaleSection504Count = F,
  Grade11FemaleWhiteCount = F,
  Grade11MaleAlaskanCount = F,
  Grade11MaleAsianCount = F,
  Grade11MaleBlackCount = F,
  Grade11MaleHawaiianCount = F,
  Grade11MaleHispanicCount = F,
  Grade11MaleIDEACount = F,
  Grade11MaleLEPCount = F,
  Grade11MaleMultiRaceCount = F,
  Grade11MaleSection504Count = F,
  Grade11MaleWhiteCount = F,
  Grade12FemaleAlaskanCount = F,
  Grade12FemaleAsianCount = F,
  Grade12FemaleBlackCount = F,
  Grade12FemaleHawaiianCount = F,
  Grade12FemaleHispanicCount = F,
  Grade12FemaleIDEACount = F,
  Grade12FemaleLEPCount = F,
  Grade12FemaleMultiRaceCount = F,
  Grade12FemaleSection504Count = F,
  Grade12FemaleWhiteCount = F,
  Grade12MaleAlaskanCount = F,
  Grade12MaleAsianCount = F,
  Grade12MaleBlackCount = F,
  Grade12MaleHawaiianCount = F,
  Grade12MaleHispanicCount = F,
  Grade12MaleIDEACount = F,
  Grade12MaleLEPCount = F,
  Grade12MaleMultiRaceCount = F,
  Grade12MaleSection504Count = F,
  Grade12MaleWhiteCount = F,
  KindergartenFemaleAlaskanCount = F,
  KindergartenFemaleAsianCount = F,
  KindergartenFemaleBlackCount = F,
  KindergartenFemaleHawaiianCount = F,
  KindergartenFemaleHispanicCount = F,
  KindergartenFemaleIDEACount = F,
  KindergartenFemaleLEPCount = F,
  KindergartenFemaleMultiRaceCount = F,
  KindergartenFemaleSection504Count = F,
  KindergartenFemaleWhiteCount = F,
  KindergartenMaleAlaskanCount = F,
  KindergartenMaleAsianCount = F,
  KindergartenMaleBlackCount = F,
  KindergartenMaleHawaiianCount = F,
  KindergartenMaleHispanicCount = F,
  KindergartenMaleIDEACount = F,
  KindergartenMaleLEPCount = F,
  KindergartenMaleMultiRaceCount = F,
  KindergartenMaleSection504Count = F,
  KindergartenMaleWhiteCount = F,
  ModifiedTime = F,
  UserIDCreatedBy = F,
  UserIDModifiedBy = F,
  fieldPaths = NULL,
  entityId = 1,
  query = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

CRDCSchoolRetentionCountV1ID
The ID of the CRDCSchoolRetentionCountV1 to return.

fieldPaths
Fields from other objects with 'Many to One' or 'One to One' relationships to the given object listed as text. Run getSchemaForObjects('CRDCSchoolRetentionCountV1') to get more field paths.

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

query
Queries to be added to the request. Ex. list(SchoolYearID = 1). Defaults to NULL.

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
A TRUE or FALSE value determining whether or not to return the field for the given CRDCSchoolRetentionCountV1. Defaults to FALSE for all return fields which, for convenience, returns all fields for the CRDCSchoolRetentionCountV1.

Value

A dataframe or of CRDCSchoolRetentionCountV1 Skyward's Knowledge Hub