Get an EmployeeFixedLengthFileFormat

This function returns a dataframe or json object of an EmployeeFixedLengthFileFormat

getEmployeeFixedLengthFileFormat(
  EmployeeFixedLengthFileFormatID,
  SkywardID = F,
  SkywardHash = F,
  EmployeeThirdPartyFormatID = F,
  NumberOfHeaderRows = F,
  NamePrefixStartPosition = F,
  NamePrefixLength = F,
  LastNameStartPosition = F,
  LastNameLength = F,
  FirstNameStartPosition = F,
  FirstNameLength = F,
  MiddleNameStartPosition = F,
  MiddleNameLength = F,
  NameSuffixStartPosition = F,
  NameSuffixLength = F,
  SocialSecurityNumberStartPosition = F,
  SocialSecurityNumberLength = F,
  FederalEINStartPosition = F,
  FederalEINLength = F,
  BirthDateStartPosition = F,
  BirthDateLength = F,
  GenderStartPosition = F,
  GenderLength = F,
  PhoneNumberStartPosition = F,
  PhoneNumberLength = F,
  PhoneTypeStartPosition = F,
  PhoneTypeLength = F,
  ExtensionStartPosition = F,
  ExtensionLength = F,
  EmailAddressStartPosition = F,
  EmailAddressLength = F,
  EmailTypeStartPosition = F,
  EmailTypeLength = F,
  AddressTypeCodeStartPosition = F,
  AddressTypeCodeLength = F,
  StreetNumberStartPosition = F,
  StreetNumberLength = F,
  DirectionalCodeStartPosition = F,
  DirectionalCodeLength = F,
  StreetNameStartPosition = F,
  StreetNameLength = F,
  AddressSecondaryUnitStartPosition = F,
  AddressSecondaryUnitLength = F,
  SecondaryUnitNumberStartPosition = F,
  SecondaryUnitNumberLength = F,
  AddressLineTwoStartPosition = F,
  AddressLineTwoLength = F,
  CityStartPosition = F,
  CityLength = F,
  StateStartPosition = F,
  StateLength = F,
  ZipCodeStartPosition = F,
  ZipCodeLength = F,
  CheckLocationStartPosition = F,
  CheckLocationLength = F,
  EmployeeNumberStartPosition = F,
  EmployeeNumberLength = F,
  StartDateStartPosition = F,
  StartDateLength = F,
  HireDateStartPosition = F,
  HireDateLength = F,
  I9DateStartPosition = F,
  I9DateLength = F,
  W4DateStartPosition = F,
  W4DateLength = F,
  VendorNumberStartPosition = F,
  VendorNumberLength = F,
  UsernameStartPosition = F,
  UsernameLength = F,
  AllowEmployeeAccessStartPosition = F,
  AllowEmployeeAccessLength = F,
  UserIDCreatedBy = F,
  CreatedTime = F,
  UserIDModifiedBy = F,
  ModifiedTime = F,
  IsAlaskanStartPosition = F,
  IsAlaskanLength = F,
  IsAsianStartPosition = F,
  IsAsianLength = F,
  IsBlackStartPosition = F,
  IsBlackLength = F,
  IsHawaiianStartPosition = F,
  IsHawaiianLength = F,
  IsHispanicStartPosition = F,
  IsHispanicLength = F,
  IsWhiteStartPosition = F,
  IsWhiteLength = F,
  ZipCodeAddOnStartPosition = F,
  ZipCodeAddOnLength = F,
  PhoneNumber2StartPosition = F,
  PhoneNumber2Length = F,
  PhoneType2StartPosition = F,
  PhoneType2Length = F,
  Extension2StartPosition = F,
  Extension2Length = F,
  EmailAddress2StartPosition = F,
  EmailAddress2Length = F,
  EmailType2StartPosition = F,
  EmailType2Length = F,
  PhoneNumber3StartPosition = F,
  PhoneNumber3Length = F,
  PhoneType3StartPosition = F,
  PhoneType3Length = F,
  Extension3StartPosition = F,
  Extension3Length = F,
  EmailAddress3StartPosition = F,
  EmailAddress3Length = F,
  EmailType3StartPosition = F,
  EmailType3Length = F,
  fieldPaths = NULL,
  entityId = 1,
  schoolYearId = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

EmployeeFixedLengthFileFormatID
The ID of the EmployeeFixedLengthFileFormat to return.

fieldPaths
Fields from other objects with 'Many to One' or 'One to One' relationships to the given object listed as text. Run getSchemaForObjects('EmployeeFixedLengthFileFormat') to get more field paths.

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
A TRUE or FALSE value determining whether or not to return the field for the given EmployeeFixedLengthFileFormat. Defaults to FALSE for all return fields which, for convenience, returns all fields for the EmployeeFixedLengthFileFormat.

Value

A dataframe or of EmployeeFixedLengthFileFormat Skyward's Knowledge Hub