Get a ReportQueue

This function returns a dataframe or json object of a ReportQueue

getReportQueue(
  ReportQueueID,
  ReportID = F,
  ReportName = F,
  ReportType = F,
  SkywardReportSystemVersion = F,
  IsFromPublishedVersion = F,
  PageCountWhenCompleted = F,
  PageCount = F,
  LogID = F,
  SaveUntil = F,
  HasContent = F,
  HasContentViewable = F,
  HasContentDownloadable = F,
  StatusCode = F,
  Status = F,
  QueueCode = F,
  Queue = F,
  CanCancel = F,
  EntityID = F,
  EntityIDList = F,
  FiscalYearID = F,
  SchoolYearID = F,
  SectionID = F,
  SchoolYearNumericYearOrCurrent = F,
  SchoolYearIDSelectedOrCurrent = F,
  FiscalYearIDSelectedOrCurrent = F,
  Application = F,
  Hostname = F,
  ThreadName = F,
  ProcessID = F,
  ReferrerPath = F,
  MediaID = F,
  CrystalParameterData = F,
  ScheduledReportID = F,
  QueryStartTime = F,
  RenderingStartTime = F,
  EndTime = F,
  QueuedDuration = F,
  QueryDuration = F,
  RenderingDuration = F,
  CachedEntity = F,
  CachedSchoolYear = F,
  CachedFiscalYear = F,
  UserIDCreatedBy = F,
  CreatedTime = F,
  UserIDModifiedBy = F,
  ModifiedTime = F,
  CurrentUserCanView = F,
  ReportFileExtensionOverride = F,
  IsCrystalReport = F,
  LastActivity = F,
  StatusAction = F,
  PromptTemplateID = F,
  FTPResultID = F,
  DataSource = F,
  MediaIDPrint = F,
  MediaIDDownload = F,
  MediaIDCsv = F,
  EncodingType = F,
  Encoding = F,
  BurstObjectIDs = F,
  HasPrintAction = F,
  HideFiscalYear = F,
  HideSchoolYear = F,
  ReportQueueIDSummaryReportSource = F,
  ReportQueueIDSummaryReport = F,
  IsViewableReport = F,
  ResultStatus = F,
  fieldPaths = NULL,
  entityId = 1,
  query = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

ReportQueueID
The ID of the ReportQueue to return.

fieldPaths
Fields from other objects with 'Many to One' or 'One to One' relationships to the given object listed as text. Run getSchemaForObjects('ReportQueue') to get more field paths.

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
A TRUE or FALSE value determining whether or not to return the field for the given ReportQueue. Defaults to FALSE for all return fields which, for convenience, returns all fields for the ReportQueue.

Value

A dataframe or of ReportQueue Skyward's Knowledge Hub