Get a W2YearEmployeeV1

This function returns a dataframe or json object of a W2YearEmployeeV1

getW2YearEmployeeV1(
  W2YearEmployeeV1ID,
  W2YearID = F,
  EmployeeID = F,
  AddressID = F,
  LocationAddressPrinted = F,
  DeliveryAddressPrinted = F,
  CityPrinted = F,
  StateID = F,
  StateCode = F,
  ZipCode = F,
  ZipCodeExtension = F,
  LocationAddressElectronicFile = F,
  DeliveryAddressElectronicFile = F,
  CityElectronicFile = F,
  EmployeeNamePrinted = F,
  FirstNameElectronicFile = F,
  MiddleNameElectronicFile = F,
  LastNameElectronicFile = F,
  SuffixElectronicFile = F,
  EmployeeSocialSecurityNumber = F,
  MaskedEmployeeSocialSecurityNumber = F,
  Box1FederalWages = F,
  Box1FederalWagesExtracted = F,
  Box2FederalIncomeTaxWithheld = F,
  Box2FederalIncomeTaxWithheldExtracted = F,
  Box3SocialSecurityWages = F,
  Box3SocialSecurityWagesExtracted = F,
  Box4SocialSecurityTaxWithheld = F,
  Box4SocialSecurityTaxWithheldExtracted = F,
  Box5MedicareWages = F,
  Box5MedicareWagesExtracted = F,
  Box6MedicareTaxWithheld = F,
  Box6MedicareTaxWithheldExtracted = F,
  Box7SocialSecurityTips = F,
  Box9AdvanceEmployeeIncomeCredit = F,
  Box10DependentCareBenefits = F,
  Box10DependentCareBenefitsExtracted = F,
  Box11NonQualifiedPlansSection457b = F,
  Box11NonQualifiedPlansSection457bExtracted = F,
  Box11NonQualifiedPlansNonSection457b = F,
  Box11NonQualifiedPlansNonSection457bExtracted = F,
  IsBox13StatutoryEmployee = F,
  IsBox13RetirementPlan = F,
  IsBox13ThirdPartySickPay = F,
  MediaIDFullSSN = F,
  MediaIDMaskedSSN = F,
  XMLData = F,
  FatalExceptionCount = F,
  NonFatalExceptionCount = F,
  UserIDCreatedBy = F,
  CreatedTime = F,
  UserIDModifiedBy = F,
  ModifiedTime = F,
  HasZeroWages = F,
  Box9VerificationCode = F,
  InternationalProvinceID = F,
  InternationalStateOrProvinceCode = F,
  AddressType = F,
  PostalCode = F,
  CountryCode = F,
  W2CountryID = F,
  fieldPaths = NULL,
  entityId = 1,
  query = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

W2YearEmployeeV1ID
The ID of the W2YearEmployeeV1 to return.

fieldPaths
Fields from other objects with 'Many to One' or 'One to One' relationships to the given object listed as text. Run getSchemaForObjects('W2YearEmployeeV1') to get more field paths.

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
A TRUE or FALSE value determining whether or not to return the field for the given W2YearEmployeeV1. Defaults to FALSE for all return fields which, for convenience, returns all fields for the W2YearEmployeeV1.

Value

A dataframe or of W2YearEmployeeV1 Skyward's Knowledge Hub