List Students

This function returns a dataframe or json object of Students

listStudents(
  searchConditionsList = NULL,
  StudentMNID = F,
  StudentStateID = F,
  StateEthnicityRaceCodeMNID = F,
  HasSection504Indicator = F,
  HasActiveSection504 = F,
  StudentID = F,
  NameID = F,
  FullNameLFM = F,
  FullNameFL = F,
  FullNameFML = F,
  AllowMediaDistribution = F,
  AllowTripsDistribution = F,
  AllowDistrictDistribution = F,
  AllowHigherEdDistribution = F,
  AllowLocalDistribution = F,
  AllowMilitaryDistribution = F,
  AllowPublicDistribution = F,
  AllowVendorsDistribution = F,
  ConversionKey = F,
  IsGraduated = F,
  StudentNumber = F,
  GraduationDate = F,
  GradYear = F,
  GraduationRequirementYear = F,
  IsCurrentActive = F,
  IsActiveAsOfDate = F,
  SpecifiedEntityYearNoShow = F,
  LanguageIDNative = F,
  EdFiCountryIDBirth = F,
  EarnedCredits = F,
  EarnedCreditsPossible = F,
  FailedCredits = F,
  EntityConfigFailedCredits = F,
  EntityConfigEarnedCredits = F,
  HasActiveAlert = F,
  HasActiveCriticalAlert = F,
  HasActiveHealthCondition = F,
  MMRStatus = F,
  HasActiveIHP = F,
  HasActiveGeneralNote = F,
  HasActiveParentalConsentNote = F,
  HealthProfessionalIDPrimaryPhysician = F,
  IsHealthProfessionalPrimaryPhysician = F,
  NameIDHospital = F,
  HealthProfessionalIDDentist = F,
  IsHealthProfessionalDentist = F,
  NameIDDentalPractice = F,
  NameIDHealthInsuranceCompany = F,
  HealthPolicyNumber = F,
  NameIDDentalInsuranceCompany = F,
  DentalPolicyNumber = F,
  HasAlertIndicator = F,
  HasCriticalAlertIndicator = F,
  HasHealthIndicator = F,
  HasIHPIndicator = F,
  HasGeneralNoteIndicator = F,
  HasParentalConsentNoteIndicator = F,
  CalculatedGradYear = F,
  CurrentDefaultEntityIsActive = F,
  CalculatedEntityYearIsActive = F,
  CalculatedGrade = F,
  GradeNumeric = F,
  Grade = F,
  Location = F,
  IsIDEASnapshot = F,
  IsIDEA = F,
  IsLEPSnapshot = F,
  IsLEP = F,
  IsGiftedTalentedSnapshot = F,
  IsSection504 = F,
  IsSection504Snapshot = F,
  IsChronicallyAbsent = F,
  HasTakenACT0912 = F,
  HasTakenSAT0912 = F,
  CorporalPunishmentIncidents = F,
  InSchoolSuspensionCount = F,
  OutOfSchoolSuspensionCount = F,
  ExpulsionWithServicesCount = F,
  ExpulsionWithoutServicesCount = F,
  ZeroToleranceWithServicesCount = F,
  ZeroToleranceWithoutServicesCount = F,
  LawEnforcementReferralCount = F,
  ArrestCount = F,
  OutOfSchoolSuspensionMissedDays = F,
  IsFederalDualEnrollment = F,
  IsFederalDistanceEducation = F,
  HasAlgebraIGrade07 = F,
  HasAlgebraIGrade08 = F,
  HasAlgebraIGrade0910 = F,
  HasAlgebraIGrade1112 = F,
  HasAlgebraIIGrade0912 = F,
  HasAdvancedMathGrade0912 = F,
  HasCalculusGrade0912 = F,
  HasGeometryGrade08 = F,
  HasGeometryGrade0912 = F,
  HasBiologyGrade0912 = F,
  HasChemistryGrade0912 = F,
  HasPhysicsGrade0912 = F,
  HasAPCourseGrade0912 = F,
  HasAPMathGrade0912 = F,
  HasAPScienceGrade0912 = F,
  HasAPOtherGrade0912 = F,
  HasPassedAlgebraIGrade07 = F,
  HasPassedAlgebraIGrade08 = F,
  HasPassedAlgebraIGrade0910 = F,
  HasPassedAlgebraIGrade1112 = F,
  UserIDCreatedBy = F,
  CreatedTime = F,
  UserIDModifiedBy = F,
  ModifiedTime = F,
  IndicatorsXML = F,
  HasActiveStudentGuardianRestrictedAccess = F,
  FirstName = F,
  MiddleName = F,
  LastName = F,
  LanguageIDPrimary = F,
  HasStudentGuardianWithAllowFamilyAccessButNoSecurity = F,
  HasStudentEntityYearForCurrentSchoolYear = F,
  HasAllowStudentAccessButNoSecurity = F,
  EarliestDistrictEnrollmentDate = F,
  AllowStudentAccess = F,
  OverrideFeeOnlinePaymentAccess = F,
  FeeOnlinePaymentOverrideType = F,
  OverrideFoodServiceOnlinePaymentAccess = F,
  FoodServiceOnlinePaymentOverrideType = F,
  DateOfFirstEntryWithdrawalInEntity = F,
  MedicaidNumber = F,
  SchoolPathExpectedSchoolName = F,
  IsIBDiplomaProgramme = F,
  FormattedVaccinationDates = F,
  SchoolPathExpectedSchoolCode = F,
  CalculatedStudentStateID = F,
  HasDuplicateStudentNumber = F,
  IsCurrentlyTransported = F,
  SpecifiedCohortNumericSchoolYear = F,
  SchoolYearIDSpecifiedCohort = F,
  StudentRankSort = F,
  HasComputerScienceGrade0912 = F,
  HasAPComputerScienceGrade0912 = F,
  CalculatedPrimaryFormattedPhoneNumber = F,
  HasCreditRecovery = F,
  SingleSexAthleticCount = F,
  ReportedBulliedSex = F,
  ReportedBulliedRace = F,
  ReportedBulliedDisability = F,
  DisciplinedBullyingSex = F,
  DisciplinedBullyingRace = F,
  DisciplinedBullyingDisability = F,
  MechanicalRestraintCount = F,
  PhysicalRestraintCount = F,
  SeclusionCount = F,
  TransferToAlternativeSchool = F,
  EntryWithdrawalIDDefaultEntityToday = F,
  AttachmentCount = F,
  LocationDateToUse = F,
  LocationEntityID = F,
  LocationSchoolYearID = F,
  HasTakenAPExam0912 = F,
  CurrentEconomicIndicator = F,
  IsStateReportingUnknownGender = F,
  HasOneNormalEnrollmentInEntityDuringSchoolYear = F,
  TotalCommunityServiceHours = F,
  HasActiveInterventionPlan = F,
  HasInterventionPlanIndicator = F,
  InSpecifiedDirectCertificationImport = F,
  EligibilityCategoryForDirectCertificationImport = F,
  EffectiveDateForDirectCertificationImport = F,
  HasActiveMethodOfInstruction = F,
  HasMethodOfInstructionIndicator = F,
  HasSpecialEducationIndicator = F,
  HasActiveSpecialEducation = F,
  CountryIDBirth = F,
  fieldPaths = NULL,
  searchConditionsGroupType = "And",
  searchSortFieldNamesList = NULL,
  searchSortFieldNamesDescendingList = NULL,
  entityId = 1,
  schoolYearId = NULL,
  page = 1,
  pageSize = 1e+05,
  flatten = T,
  returnResponse = F
)

Arguments

searchConditionsList
A list of search conditions to filter results which are joined by the searchConditionsGroupType. Of the form FieldName ConditionType SearchCondition. For example, c('StudentID LessEqual 500', 'LastName Like Ander%'). Run listSearchConditionTypes for a list of ConditionTypes. Defaults to NULL (unfiltered).

fieldPaths
Fields from other objects with 'Many to One' or 'One to One' relationships to the given object listed as text. Run getSchemaForObjects('Student') to get more field paths.

searchConditionsGroupType
The conjunction which joins multiple searchConditions in the searchConditionsList. Either 'Or' or 'And'. Defaults to 'And'.

searchSortFieldNamesList
The list of fields sort results by. Defaults to NULL (unsorted).

searchSortFieldNamesDescendingList
A list of T/F values corresponding to whether to sort each field in searchSortFieldNamesList in descending order. Defaults to F for each FieldName in searchSortFieldNamesList.

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

page
Results are paginated. The page of results to return. Default is 1.

pageSize
Results are paginated. The number of records per page to return. Default is 100,000 (essentially all records for most objects).

flatten
Whether to flatten results into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
A TRUE or FALSE value determining whether or not to return the field for the given Students. Defaults to FALSE for all return fields which, for convenience, returns all fields for the Students.

Value

A list of Students Skyward's Knowledge Hub