Modify a CRDCSchoolNonIDEARestraintV1

This function modifies a CRDCSchoolNonIDEARestraintV1

modifyCRDCSchoolNonIDEARestraintV1(
  CRDCSchoolNonIDEARestraintV1ID,
  CivilRightsDataCollectionID = NULL,
  CivilRightsDataCollectionRunHistoryID = NULL,
  FederalNCESSchoolID = NULL,
  MechanicalFemaleAlaskanCount = NULL,
  MechanicalFemaleAsianCount = NULL,
  MechanicalFemaleBlackCount = NULL,
  MechanicalFemaleHawaiianCount = NULL,
  MechanicalFemaleHispanicCount = NULL,
  MechanicalFemaleLEPCount = NULL,
  MechanicalFemaleMultiRaceCount = NULL,
  MechanicalFemaleSection504Count = NULL,
  MechanicalFemaleWhiteCount = NULL,
  MechanicalMaleAlaskanCount = NULL,
  MechanicalMaleAsianCount = NULL,
  MechanicalMaleBlackCount = NULL,
  MechanicalMaleHawaiianCount = NULL,
  MechanicalMaleHispanicCount = NULL,
  MechanicalMaleLEPCount = NULL,
  MechanicalMaleMultiRaceCount = NULL,
  MechanicalMaleSection504Count = NULL,
  MechanicalMaleWhiteCount = NULL,
  PhysicalFemaleAlaskanCount = NULL,
  PhysicalFemaleAsianCount = NULL,
  PhysicalFemaleBlackCount = NULL,
  PhysicalFemaleHawaiianCount = NULL,
  PhysicalFemaleHispanicCount = NULL,
  PhysicalFemaleLEPCount = NULL,
  PhysicalFemaleMultiRaceCount = NULL,
  PhysicalFemaleSection504Count = NULL,
  PhysicalFemaleWhiteCount = NULL,
  PhysicalMaleAlaskanCount = NULL,
  PhysicalMaleAsianCount = NULL,
  PhysicalMaleBlackCount = NULL,
  PhysicalMaleHawaiianCount = NULL,
  PhysicalMaleHispanicCount = NULL,
  PhysicalMaleLEPCount = NULL,
  PhysicalMaleMultiRaceCount = NULL,
  PhysicalMaleSection504Count = NULL,
  PhysicalMaleWhiteCount = NULL,
  SeclusionFemaleAlaskanCount = NULL,
  SeclusionFemaleAsianCount = NULL,
  SeclusionFemaleBlackCount = NULL,
  SeclusionFemaleHawaiianCount = NULL,
  SeclusionFemaleHispanicCount = NULL,
  SeclusionFemaleLEPCount = NULL,
  SeclusionFemaleMultiRaceCount = NULL,
  SeclusionFemaleSection504Count = NULL,
  SeclusionFemaleWhiteCount = NULL,
  SeclusionMaleAlaskanCount = NULL,
  SeclusionMaleAsianCount = NULL,
  SeclusionMaleBlackCount = NULL,
  SeclusionMaleHawaiianCount = NULL,
  SeclusionMaleHispanicCount = NULL,
  SeclusionMaleLEPCount = NULL,
  SeclusionMaleMultiRaceCount = NULL,
  SeclusionMaleSection504Count = NULL,
  SeclusionMaleWhiteCount = NULL,
  entityId = 1,
  query = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

query
Queries to be added to the request. Ex. list(SchoolYearID = 1). Defaults to NULL.

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the modified CRDCSchoolNonIDEARestraintV1. Each defaults to NULL.

Value

The modified CRDCSchoolNonIDEARestraintV1 Skyward's Knowledge Hub