Modify a CreditCardTransactionFixedLengthFileFormat

This function modifies a CreditCardTransactionFixedLengthFileFormat

modifyCreditCardTransactionFixedLengthFileFormat(
  CreditCardTransactionFixedLengthFileFormatID,
  CreditCardTransactionThirdPartyFormatID = NULL,
  NumberOfHeaderRows = NULL,
  TransactionImportIDStartPosition = NULL,
  SequenceNumberStartPosition = NULL,
  UniqueCardIdentifierStartPosition = NULL,
  TransactionDateStartPosition = NULL,
  TransactionAmountStartPosition = NULL,
  SplitAmountStartPosition = NULL,
  GeneralDefaultCodeOneStartPosition = NULL,
  GeneralDefaultCodeTwoStartPosition = NULL,
  GeneralDefaultCodeThreeStartPosition = NULL,
  GeneralDefaultCodeFourStartPosition = NULL,
  GeneralDefaultCodeFiveStartPosition = NULL,
  GeneralDefaultCodeSixStartPosition = NULL,
  GeneralDefaultCodeSevenStartPosition = NULL,
  GeneralDefaultCodeEightStartPosition = NULL,
  GeneralDefaultCodeNineStartPosition = NULL,
  GeneralDefaultCodeTenStartPosition = NULL,
  GeneralDefaultCodeElevenStartPosition = NULL,
  GeneralDefaultCodeTwelveStartPosition = NULL,
  VendorNameStartPosition = NULL,
  VendorCityStartPosition = NULL,
  VendorStateProvinceStartPosition = NULL,
  VendorZipCodeStartPosition = NULL,
  VendorCountryStartPosition = NULL,
  TransactionDescriptionStartPosition = NULL,
  SplitDescriptionStartPosition = NULL,
  TransactionImportIDLength = NULL,
  SequenceNumberLength = NULL,
  UniqueCardIdentifierLength = NULL,
  TransactionDateLength = NULL,
  TransactionAmountLength = NULL,
  SplitAmountLength = NULL,
  GeneralDefaultCodeOneLength = NULL,
  GeneralDefaultCodeTwoLength = NULL,
  GeneralDefaultCodeThreeLength = NULL,
  GeneralDefaultCodeFourLength = NULL,
  GeneralDefaultCodeFiveLength = NULL,
  GeneralDefaultCodeSixLength = NULL,
  GeneralDefaultCodeSevenLength = NULL,
  GeneralDefaultCodeEightLength = NULL,
  GeneralDefaultCodeNineLength = NULL,
  GeneralDefaultCodeTenLength = NULL,
  GeneralDefaultCodeElevenLength = NULL,
  GeneralDefaultCodeTwelveLength = NULL,
  VendorNameLength = NULL,
  VendorCityLength = NULL,
  VendorStateProvinceLength = NULL,
  VendorZipCodeLength = NULL,
  VendorCountryLength = NULL,
  TransactionDescriptionLength = NULL,
  SplitDescriptionLength = NULL,
  UniqueAccountIdentifierStartPosition = NULL,
  UniqueAccountIdentifierLength = NULL,
  entityId = 1,
  schoolYearId = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the modified CreditCardTransactionFixedLengthFileFormat. Each defaults to NULL.

Value

The modified CreditCardTransactionFixedLengthFileFormat Skyward's Knowledge Hub