Modify a NameMedicationSecuredNote

This function modifies a NameMedicationSecuredNote

modifyNameMedicationSecuredNote(
  NameMedicationSecuredNoteID,
  NameMedicationID = NULL,
  Note = NULL,
  entityId = 1,
  schoolYearId = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

schoolYearId
The id of the schoolYear. Run listSchoolYears for a list of school years. Defaults to NULL (all school years).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the modified NameMedicationSecuredNote. Each defaults to NULL.

Value

The modified NameMedicationSecuredNote Skyward's Knowledge Hub