Modify a Student

This function modifies a Student

modifyStudent(
  StudentID,
  AllowDistrictDistribution = NULL,
  AllowHigherEdDistribution = NULL,
  AllowLocalDistribution = NULL,
  AllowMediaDistribution = NULL,
  AllowMilitaryDistribution = NULL,
  AllowPublicDistribution = NULL,
  AllowStudentAccess = NULL,
  AllowTripsDistribution = NULL,
  AllowVendorsDistribution = NULL,
  ConversionKey = NULL,
  CountryIDBirth = NULL,
  DentalPolicyNumber = NULL,
  EdFiCountryIDBirth = NULL,
  FeeOnlinePaymentOverrideType = NULL,
  FoodServiceOnlinePaymentOverrideType = NULL,
  Grade = NULL,
  GradeNumeric = NULL,
  GraduationDate = NULL,
  GraduationRequirementYear = NULL,
  GradYear = NULL,
  HasAlertIndicator = NULL,
  HasCriticalAlertIndicator = NULL,
  HasGeneralNoteIndicator = NULL,
  HasHealthIndicator = NULL,
  HasIHPIndicator = NULL,
  HasInterventionPlanIndicator = NULL,
  HasMethodOfInstructionIndicator = NULL,
  HasParentalConsentNoteIndicator = NULL,
  HasSection504Indicator = NULL,
  HasSpecialEducationIndicator = NULL,
  HealthPolicyNumber = NULL,
  HealthProfessionalIDDentist = NULL,
  HealthProfessionalIDPrimaryPhysician = NULL,
  IndicatorsXML = NULL,
  IsGraduated = NULL,
  IsIBDiplomaProgramme = NULL,
  IsStateReportingUnknownGender = NULL,
  LanguageIDNative = NULL,
  LanguageIDPrimary = NULL,
  MedicaidNumber = NULL,
  NameID = NULL,
  NameIDDentalInsuranceCompany = NULL,
  NameIDDentalPractice = NULL,
  NameIDHealthInsuranceCompany = NULL,
  NameIDHospital = NULL,
  OverrideFeeOnlinePaymentAccess = NULL,
  OverrideFoodServiceOnlinePaymentAccess = NULL,
  StateEthnicityRaceCodeMNID = NULL,
  StudentNumber = NULL,
  StudentStateID = NULL,
  entityId = 1,
  query = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

query
Queries to be added to the request. Ex. list(SchoolYearID = 1). Defaults to NULL.

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the modified Student. Each defaults to NULL.

Value

The modified Student Skyward's Knowledge Hub