Modify an UserMessageSetting

This function modifies an UserMessageSetting

modifyUserMessageSetting(
  UserMessageSettingID,
  UserIDOwner = NULL,
  CopyMessagesToEmail = NULL,
  CopyDisciplineMessagesToEmail = NULL,
  CopyFeeManagementMessagesToEmail = NULL,
  CopyFoodServiceMessagesToEmail = NULL,
  CopyGradingMessagesToEmail = NULL,
  CopySchedulingMessagesToEmail = NULL,
  CopyAttendanceMessagesToEmail = NULL,
  CopyOnlineFormMessagesToEmail = NULL,
  EnableCompletedGradeChangeNotification = NULL,
  EnableCompletedGradeChangeNotificationEmail = NULL,
  CopyGradebookMessagesToEmail = NULL,
  GradebookHighThreshold = NULL,
  GradebookLowThreshold = NULL,
  GradebookLowAssignmentThreshold = NULL,
  AssignmentScoreLowNotification = NULL,
  AssignmentScoreLowNotificationEmail = NULL,
  GradebookHighAssignmentThreshold = NULL,
  AssignmentScoreHighNotification = NULL,
  AssignmentScoreHighNotificationEmail = NULL,
  CurrentGradeScoreLowNotification = NULL,
  CurrentGradeScoreLowNotificationEmail = NULL,
  CurrentGradeScoreHighNotification = NULL,
  CurrentGradeScoreHighNotificationEmail = NULL,
  MissingAssignmentNotification = NULL,
  MissingAssignmentNotificationEmail = NULL,
  EnableGradebookLastEntryNotificationEmail = NULL,
  EnableGradebookGradeChangeRequestDeniedEmail = NULL,
  EnableGradebookGradeChangeRequestNotificationEmail = NULL,
  EnableOnlineAssignmentAvailableNotificationEmail = NULL,
  EnableOnlineAssingmentScoresAvailableNotificationEmail = NULL,
  CopyEnrollmentMessagesToEmail = NULL,
  CopyGraduationRequirementsMessagesToEmail = NULL,
  EnableCompletedCareerPlanChangeNotification = NULL,
  EnableCompletedCareerPlanChangeNotificationEmail = NULL,
  OnlySendAssignmentScoreLowNotificationsOncePerAssignment = NULL,
  OnlySendAssignmentScoreHighNotificationsOncePerAssignment = NULL,
  OnlySendCurrentGradeScoreLowNotificationsOnce = NULL,
  OnlySendCurrentGradeScoreHighNotificationsOnce = NULL,
  OnlySendMissingAssignmentNotificationForCurrentGradingPeriod = NULL,
  EnableStudentScheduleChangeNotification = NULL,
  EnableStudentScheduleChangeNotificationEmail = NULL,
  OnlySendMissingAssignmentNotificationsOncePerAssignment = NULL,
  CopyFamilyAccessMessagesToEmail = NULL,
  CopyMTSSMessagesToEmail = NULL,
  entityId = 1,
  query = NULL,
  flatten = T,
  returnResponse = F
)

Arguments

entityId
The id of the entity (school). Run listEntities for a list of entities. Defaults to 1 (district).

flatten
Whether to flatten result into a dataframe or return the json object. Default is TRUE (flatten to dataframe).

returnResponse
Whether to return the server response instead of the results. Useful for debugging. Default is FALSE.

fieldNames
The field values to give the modified UserMessageSetting. Each defaults to NULL.

Value

The modified UserMessageSetting Skyward's Knowledge Hub